requirements.txt
flask
flask-smorest
python-dotenv
marshmallow
flask_sqlalchemy
psycopg2